migrate spectral modifiers to pvlib.spectrum
