+++ pvlib/__init__.py
@@ -1,6 +1,9 @@
 from pvlib.version import __version__  # noqa: F401
 
 from pvlib import (  # noqa: F401
+    # list spectrum first so it's available for atmosphere & pvsystem (GH 1628)
+    spectrum,
+
     atmosphere,
     bifacial,
     clearsky,
@@ -20,7 +23,6 @@ from pvlib import (  # noqa: F401
     soiling,
     solarposition,
     spa,
-    spectrum,
     temperature,
     tools,
     tracking,
