+++ pvlib/spectrum/__init__.py
@@ -1,3 +1,8 @@
 from pvlib.spectrum.spectrl2 import spectrl2  # noqa: F401
-from pvlib.spectrum.mismatch import (get_example_spectral_response, get_am15g,
-                                     calc_spectral_mismatch_field)
+from pvlib.spectrum.mismatch import (
+    calc_spectral_mismatch_field,
+    first_solar_spectral_correction,
+    get_am15g,
+    get_example_spectral_response,
+    sapm_spectral_correction,
+)  # noqa: F401
